backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
pymysql==1.1.2
pydantic-settings==2.10.1
email-validator==2.2.0
python-jose[cryptography]==3.5.0
cryptography==44.0.2
bcrypt==4.2.0
python-multipart==0.0.20
httpx==0.28.1
openpyxl==3.1.5
